+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.1.1-cf] - 2019-11-07
 ### Fixed
 - Fixed missing argument in [`cloudflare.py`](cloudflare.py) `update_subdomain`
+- Fixed `config['domain']` -> `config.DOMAIN`
 
 ## [1.1-cf] - 2019-10-22
 ### Added